services/requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
finnhub-python==2.4.0
idna==2.10
psycopg2-binary==2.8.6
pudb==2020.1
Pygments==2.8.1
python-dateutil==2.8.1
python-dotenv==0.15.0
requests==2.25.1
six==1.15.0
soupsieve==2.2.1
urllib3==1.26.4
urwid==2.1.2
websocket-client==0.58.0
